# src/main/resources/static/openapi.yml
openapi: 3.1.0
info:
  title: Survey Engine MVP API
  version: 1.0.0
  description: |
    Comprehensive API contract for the Survey Engine MVP.

    This specification is written for product, implementation, and integration teams.
    Every endpoint description intentionally includes:
    1) why the endpoint is needed,
    2) what the endpoint does,
    3) how the endpoint works inside the platform lifecycle.

    Security model summary:
    - Admin APIs accept engine-issued JWT via HttpOnly access_token cookie (primary) or Bearer header (fallback).
    - Admin sessions are maintained with refresh-token cookie rotation.
    - Responder submission endpoint is public at transport level, but private campaigns enforce tenant-scoped responder authentication in business logic.
servers:
  - url: /
    description: Same-origin deployment
...
security:
  - bearerAuth: []
paths:
  /api/v1/admin/auth/register:
    post:
      tags: [Admin Authentication]
      operationId: registerAdmin
      security: []
      summary: Register a tenant admin account and bootstrap tenant access
      description: |
        Why this endpoint is needed:
        A brand-new subscriber needs a secure, first-class admin identity to operate the platform. Without this endpoint,
        onboarding would require manual platform-team intervention and would break self-service SaaS adoption.

        What this endpoint does:
        It creates the first admin user for a tenant, ensures tenant provisioning exists, ensures a trial subscription exists,
        sets access/refresh tokens as HttpOnly cookies, and returns safe user context in response body.

        How this endpoint does it:
        The service validates uniqueness of email, provisions tenant records when necessary, enforces admin-user quota policy,
        stores password hash securely, issues an engine-signed JWT access token, creates a refresh token row, and controller
        writes both tokens into secure cookies before returning AuthUserResponse.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Admin account created and authenticated session returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/admin/auth/login:
    post:
      tags: [Admin Authentication]
      operationId: loginAdmin
      security: []
      summary: Log in an existing admin and issue fresh auth cookies
      description: |
        Why this endpoint is needed:
        Admin operators need a secure login path to start authenticated management sessions for tenant resources.

        What this endpoint does:
        It authenticates the admin using email/password, revokes prior refresh tokens for session hygiene, and issues
        a fresh access token plus a new refresh token as HttpOnly cookies.

        How this endpoint does it:
        The service verifies account status and password hash, rotates refresh-token state to avoid stale session reuse,
        and controller writes cookies while returning safe user context payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful and new auth cookies issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/auth/refresh:
    post:
      tags: [Admin Authentication]
      operationId: refreshAdminToken
      security: []
      summary: Rotate refresh cookie and issue a new admin session
      description: |
        Why this endpoint is needed:
        Access tokens are intentionally short-lived. A secure refresh mechanism is required to keep admins signed in
        without repeatedly collecting credentials.

        What this endpoint does:
        It reads refresh token from HttpOnly cookie, validates it, rotates token state, and sets fresh access/refresh cookies.

        How this endpoint does it:
        Controller extracts refresh_token cookie, service validates persistence state (including expiry/revocation),
        applies one-time-use rotation semantics, and controller rewrites secure cookies.
      responses:
        '200':
          description: Cookie rotation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/admin/auth/me:
    get:
      tags: [Admin Authentication]
      operationId: getCurrentAdmin
      summary: Retrieve current authenticated admin profile
      description: |
        Why this endpoint is needed:
        Frontend needs a lightweight session-hydration endpoint after reload to confirm authentication and tenant context.

        What this endpoint does:
        It returns current authenticated user's profile details (no tokens).

        How this endpoint does it:
        JWT filter populates security/tenant context from cookie or bearer token, and controller maps context to response DTO.
      responses:
        '200':
          description: Current admin profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Engine-issued admin access token
...
  responses:
    ValidationError:
      description: Request validation failed or business rule validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Referenced resource does not exist or is not visible in tenant scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Requested action conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    RegisterRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthUserResponse:
      type: object
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        tenantId: { type: string }
        role: { type: string, enum: [SUPER_ADMIN, ADMIN, EDITOR, VIEWER] }
...
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
              rejectedValue: {}